+++ crates/core/src/doc/changefeeds.rs
@@ -3,7 +3,6 @@ use crate::dbs::Options;
 use crate::dbs::Statement;
 use crate::doc::Document;
 use crate::err::Error;
-use crate::kvs::cache;
 
 impl Document {
 	pub async fn process_changefeeds(
@@ -19,31 +18,10 @@ impl Document {
 		// Get the NS + DB
 		let ns = opt.ns()?;
 		let db = opt.db()?;
-		// Get the transaction
-		let txn = ctx.tx();
-		// Get the table
+		// Get the table for this record
 		let tbv = self.tb(ctx, opt).await?;
-		// Get the database and the table for the record
-		let dbv = match ctx.get_cache() {
-			// A cache is present on the context
-			Some(cache) if txn.local() => {
-				// Get the cache entry key
-				let key = cache::ds::Lookup::Db(ns, db);
-				// Get or update the cache entry
-				match cache.get(&key) {
-					Some(val) => val,
-					None => {
-						let val = txn.get_or_add_db(ns, db, opt.strict).await?;
-						let val = cache::ds::Entry::Any(val.clone());
-						cache.insert(key, val.clone());
-						val
-					}
-				}
-				.try_into_type()
-			}
-			// No cache is present on the context
-			_ => txn.get_or_add_db(ns, db, opt.strict).await,
-		}?;
+		// Get the database for this record
+		let dbv = self.db(ctx, opt).await?;
 		// Get the changefeed definition on the database
 		let dbcf = dbv.as_ref().changefeed.as_ref();
 		// Get the changefeed definition on the table
@@ -52,7 +30,7 @@ impl Document {
 		if let Some(cf) = dbcf.or(tbcf) {
 			// Create the changefeed entry
 			if let Some(id) = &self.id {
-				txn.lock().await.record_change(
+				ctx.tx().lock().await.record_change(
 					ns,
 					db,
 					tbv.name.as_str(),
